+++ src/component/greeting_component.jsx
@@ -3,12 +3,12 @@ import { useRecoilState } from 'recoil'
 import icon from '../img/icon-128.png'
 import { hot } from 'react-hot-loader'
 import TestButton from './button_component'
-import { recoilPageState, recoilNetWork } from '../states/recoilPageState'
+import { recoilPageState } from '../states/recoilPageState'
 import { Page } from '../enum/enum'
 import { makeStyles } from '@material-ui/core'
 import Login from './login_component'
 import NewAccount from './new_account_component'
-import RestoreAccount from './restore_account_component'
+import RestoreAccount from './recover_account_component'
 import Account from './account_component'
 import Transfer from './transfer_component'
 import * as crypto from 'crypto-js'
